Separate methods from collection class to individual classes
