+++ .rubocop.yml
@@ -18,6 +18,8 @@ Metrics/ClassLength:
 Metrics/BlockLength:
   IgnoredMethods: ["describe"]
   Max: 30
+Metrics/CyclomaticComplexity:
+  Max: 9
 
 Style/Documentation:
   Enabled: false
