+++ main.rb
@@ -1,0 +1,4 @@
+require_relative 'modules'
+
+collection = Collection.new
+collection.run
